refactor(sidescan): extract shared decode into the library (#184)
The per-ping RawSonarImage decode was inline in mosaic_node and replicated
in the offline importer (deferred until #177/PR181 merged). Move it to a
shared marine_sidescan_mosaic/decode unit (returns double; importer casts
to float for Tier-1) so the live node and the offline pipeline use one
implementation. First step of the ADR-0006 D12 shared per-ping engine.
No behavior change; all gtest logic green (uncrustify failures are the
known local 0.78.1 version drift, flagging untouched base files too).

Part of #184
